scripts/eval/api-bank/requirements.txt
requests==2.25.1
sentence_transformers==2.2.2
tqdm==4.64.1
nltk==3.6.5
googletrans==3.1.0a0
selenium
bs4
rank_bm25
